requirements.txt
openai==1.31.1
requests==2.31.0
PyMuPDF==1.24.5
pillow==9.4.0
fastapi==0.95.2
python-dotenv==1.0.0
beautifulsoup4==4.12.2
selenium==4.22.0
lxml==5.2.2
Flask==3.0.3
gunicorn==22.0.0
flask_cors